rust-port: reduce remaining libsystemd lint debt

+++ src/libsystemd/rust/id128_util.rs
@@ -7,8 +7,8 @@ use std::str::FromStr;
 
 pub type Result<T> = std::result::Result<T, i32>;
 
-pub const NEG_EINVAL: i32 = -(libc::EINVAL as i32);
-pub const NEG_ENXIO: i32 = -(libc::ENXIO as i32);
+pub const NEG_EINVAL: i32 = -libc::EINVAL;
+pub const NEG_ENXIO: i32 = -libc::ENXIO;
 
 #[repr(C, align(8))]
 #[derive(Clone, Copy, Debug, Default, PartialEq, Eq, Hash)]
